+++ src/u1171639/sensor/main/java/monitor/WaterLevelMonitor.java
@@ -2,7 +2,7 @@ package u1171639.sensor.main.java.monitor;
 
 public interface WaterLevelMonitor {
 	public void monitorWaterLevel();
-	public float getWaterLevel();
+	public double getWaterLevel();
 	public boolean isAlarmRaised();
 	public void resumeMonitoring();
 	public void pauseMonitoring();
